pyproject.toml
[project]
name = "aviutl-whisper"
version = "0.1.0"
description = "m4a音声ファイルからWhisperで文字起こし＋話者分離を行うGUIツール"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "faster-whisper>=1.1.0",
    "speechbrain>=1.0.0",
    "pydub>=0.25.1",
    "audioop-lts>=0.2.1",
    "pywebview>=5.0",
    "scikit-learn>=1.5.0",
    "numpy>=1.26.0",
]

[project.scripts]
aviutl-whisper = "aviutl_whisper.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aviutl_whisper"]

[dependency-groups]
dev = [
    "pyinstaller>=6.0",
]
